refactor `HasMiddleware` trait
This patch refactors and removes the `HasMiddleware` trait. Instead,
we're left with two new traits, `Handle` and `WrapWith`. The former
exposes a new method, `handle`, which is a convenience method for
handling requests and abstracts away the need to map errors manually.
The latter is a re-naming of `HasMiddleware` to a perhaps more idiomatic
convention.

Note that this is a breaking change since `HasMiddleware` is removed
entirely.